fix: implement critical, high, and medium priority action items for BedrockSystemMessage
CRITICAL FIXES:
- Add streaming integration tests for BedrockSystemMessage with granular cache points
  * Test streaming with single BedrockSystemMessage (granular cache points)
  * Test streaming with multiple BedrockSystemMessage instances
- Add tests for 3+ BedrockSystemMessage instances
  * Verify 3 messages aggregate cache points correctly (at limit: 2+1+1=4)
  * Verify exceeding limit with 3 messages (2+2+2=6 exceeds limit of 4)
- Add tests combining multiple cache point placements and sources
  * Validate interaction between placement-based and granular cache points
  * Verify correct counting when placement sources conflict

HIGH PRIORITY FIXES:
- Add MAX_CONTENT_BLOCKS (10) + MAX_CACHE_POINTS (4) boundary test
  * Verify 10 content blocks with 4 cache points work together
- Add comprehensive edge case tests
  * AFTER_USER_MESSAGE without UserMessage (not counted)
  * AFTER_TOOLS counting only when placement=AFTER_TOOLS AND hasTools=true
  * Placement-only scenarios (null placement with hasTools=true)
- Add Unicode and special character handling tests
  * Multi-language Unicode characters (Chinese, Arabic, Cyrillic, Greek)
  * Emoji and special symbols in cache point content
  * Mixed Unicode with whitespace
- Add whitespace-only content block tests
  * Verify rejection of spaces-only, tabs-only, newlines-only content
  * Both with and without cache points

MEDIUM PRIORITY FIXES:
- Fix misleading comment in should_handle_after_system_with_core_system_message_and_tools
  * Updated comment to accurately reflect that AFTER_SYSTEM is NOT counted
    when last system message is BedrockSystemMessage (count: 2+0+1=3, not 4)
- Change cachePointCount() return type from long to int
  * Method now returns int directly instead of requiring cast
  * Updated toString() to use int instead of long
  * Removed unnecessary cast in AbstractBedrockChatModel.countTotalCachePoints()
- Add test for AFTER_TOOLS without tools
  * Verify AFTER_TOOLS placement ignored when hasTools=false

NEW TESTS ADDED:
- BedrockSystemMessageExtractionTest: 11 new test methods (total: 29)
  * Multiple message scenarios (critical)
  * All placement combination tests (critical)
  * Boundary and edge case tests (high priority)
- BedrockPromptCachingIT: 5 new test methods (total: 19)
  * Streaming integration tests (critical)
  * Complex cache point scenarios (high priority)
  * Placement-specific tests (AFTER_SYSTEM, AFTER_USER_MESSAGE, AFTER_TOOLS)
- BedrockSystemMessageTest: 9 new test methods (total: 42)
  * Unicode and special character tests (high priority)
  * Whitespace-only content block tests (high priority)

All 102 tests pass (20 skipped). Code properly formatted with spotless.

+++ langchain4j-bedrock/src/main/java/dev/langchain4j/model/bedrock/AbstractBedrockChatModel.java
@@ -492,7 +492,7 @@ abstract class AbstractBedrockChatModel {
             }
 
             if (message instanceof BedrockSystemMessage bedrockMsg) {
-                count += (int) bedrockMsg.cachePointCount();
+                count += bedrockMsg.cachePointCount();
                 lastSystemIsCoreMessage = false;
                 hasAnySystemMessage = true;
             } else if (message instanceof SystemMessage) {
